Fix plugin logic
Prior to this commit, the logic being used here does not match the logic
that MkDocs uses internally to generate the final site_dir.
More specifically: we don't take into account the `use_directory_urls`
setting at all, which alters how the site_dir is built (in regard to
URLs and paths).