Document redirection link
The redirection may be broken if the javascript of the page cannot be
executed because of a [middleware](https://backstage.io/docs/features/techdocs/)
or if the browser of the client disabled it. Also note that the meta
redirect is set as [deprecated by the W3C](https://www.w3.org/TR/WCAG10-HTML-TECHS/#meta-element).

With those conditions, the human may land on a page which is in fact a
dead end because no information is provided to tell him what should
have happened.

This PR provide a link to clic to follow manually the redirection.